[frontend] POST /room and GET /room/id with authentication
