+++ frontend/app/lib/client-actions.ts
@@ -30,30 +30,6 @@ export async function createUser(formData: FormData) {
   }
 }
 
-export async function createRoom(formData: FormData) {
-  const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/room`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      name: formData.get("name"),
-    }),
-  });
-  const data = await res.json();
-  if (!res.ok) {
-    toast({
-      title: res.status + " " + res.statusText,
-      description: data.message,
-    });
-  } else {
-    toast({
-      title: "Success",
-      description: "Room created successfully.",
-    });
-    redirect(`/room/${data.id}`, RedirectType.push);
-  }
-}
 
 export async function updateRoom(
   event: React.FormEvent<HTMLFormElement>,
